Notifications fixed
So it turns out that the runtime just... doesn't like a lot of parameters being passed out at once? so all of the options, except for type, have been put into one array.

the declaration for the typescipt is now as follows:
declare namespace notification {
   function send(options: string[], google: string[], apple: string[], webpush: string[], http: string[], typ_et: u32): void
}

options are in the order: universe, title, body, custom key, custom data, priority, topic

universe is required. will change depending on how exactly the system is set up.
topic is required if apple notifications are being sent.
all others are optional.